add separate component for payment reqeust table list
